railway-orchestrator/requirements.txt
# Railway Cloud Orchestrator - Python Dependencies

# Database
psycopg2-binary==2.9.9

# HTTP requests (for webhooks)
requests==2.31.0

# Utilities
python-dotenv==1.0.0
